Feature/issue 63 관리자 모듈 사용 X

+++ Match-Infrastructure/src/main/java/com/example/matchinfrastructure/pay/nice/dto/NicePaymentAuth.java
@@ -41,7 +41,7 @@ public class NicePaymentAuth {
 
     private String payMethod;
 
-    private int amount;
+    private Long amount;
 
     private int balanceAmt;
 
@@ -82,7 +82,7 @@ public class NicePaymentAuth {
     private static class Cancels{
         private String tid;
 
-        private int amount;
+        private Long amount;
 
         private String cancelledAt;
 
@@ -140,7 +140,7 @@ public class NicePaymentAuth {
 
         private String status;
 
-        private int amount;
+        private Long amount;
 
         private int taxFreeAmt;
